requirements.txt
typer
rich
pyyaml
requests
dnspython
checkdmarc
psutil
